Extracted GODialog from GOTabbedDialog
