+++ src/grandorgue/dialogs/common/GOTabbedDialog.h
@@ -10,24 +10,17 @@
 
 #include <vector>
 
-#include <wx/propdlg.h>
-
-#include "GODialogCloser.h"
+#include "GODialog.h"
 
 class wxBookCtrlBase;
 class wxPanel;
-class wxSizer;
 
 class GODialogTab;
 
-class GOTabbedDialog : public wxPropertySheetDialog, public GODialogCloser {
+class GOTabbedDialog : public GODialog {
 private:
-  const wxString m_name;
   std::vector<wxString> m_TabNames;
   wxBookCtrlBase *p_book;
-  wxSizer *p_ButtonSizer;
-
-  void OnHelp(wxCommandEvent &event);
 
 protected:
   GOTabbedDialog(
@@ -36,15 +29,14 @@ protected:
     const wxString &title, // translated
     long addStyle = 0);
 
-  wxSizer *GetButtonSizer() const { return p_ButtonSizer; }
-
   void AddTab(wxPanel *tab, const wxString &tabName, const wxString &tabTitle);
   void AddTab(GODialogTab *tab);
 
 public:
   const wxString &GetCurrTabName() const;
-
   wxBookCtrlBase *GetBook() const { return p_book; }
+  
+  const wxString &GetHelpSuffix() const override { return GetCurrTabName(); }
 
   void NavigateToTab(const wxString &tabName);
 
@@ -52,7 +44,6 @@ public:
   virtual bool Validate() override;
   virtual bool TransferDataFromWindow() override;
 
-  DECLARE_EVENT_TABLE()
 };
 
 #endif /* GOTABBEDDIALOG_H */
